Fix 64-bit thread handles during shutdown

+++ src/sys/sys_public.h
@@ -521,7 +521,7 @@ typedef enum {
 
 typedef struct {
 	const char *	name;
-	int				threadHandle;
+	uintptr_t		threadHandle;
 	unsigned long	threadId;
 } xthreadInfo;
